[Bronze III] Title: 수학은 체육과목 입니다, Time: 108 ms, Memory: 14160 KB -BaekjoonHub

+++ 백준/Bronze/15894. 수학은 체육과목 입니다/수학은 체육과목 입니다.java
@@ -4,7 +4,8 @@ import java.io.*;
 class Main{
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        long n = Long.parseLong(br.readLine());
+        long n = Integer.parseInt(br.readLine());
+        
         System.out.println(4*n);
     }
 }
